requirements.txt
deepface==0.0.79
dominant_color_detection==0.1.8
fastapi==0.103.1
imutils==0.5.4
keras==2.13.1
matplotlib==3.7.3
mediapipe==0.10.3
numpy==1.22.4
opencv_contrib_python==4.7.0.68
opencv_python==4.6.0.66
Pillow==10.0.0
PyYAML==6.0.1
scikit_image==0.19.3
scikit_learn==1.3.0
Shapely==2.0.1
starlette==0.31.1
tensorflow==2.13.0
tensorflow_intel==2.13.0
torch==1.13.1
ultralytics==8.0.138
uvicorn==0.23.2
